pregnancy_coach/app.py
from dotenv import load_dotenv
from langchain_groq import ChatGroq
from langchain_core.messages import HumanMessage, SystemMessage, AIMessage
load_dotenv()


llm = ChatGroq(
    model="llama3-70b-8192",
    temperature=0,
    max_tokens=None,
    timeout=None,
    max_retries=2,
)

system_prompt = """You are an experienced pregnancy coach with a deep understanding of pregnancy-related topics, including physical and emotional health, nutrition, prenatal care, labor, and postpartum care.
Your role is to provide accurate, empathetic, and supportive responses to pregnant women seeking advice or information specifically related to pregnancy. 
If a query falls outside the scope of pregnancy, deny answering that politely and in that case keep it short, dont answer what user has asked.
Ensure that your answers are clear, concise, and based on the latest medical guidelines. Use a friendly and reassuring tone.

Example user queries:

"What foods should I avoid during pregnancy?"
"How can I manage morning sickness?"
"What are the signs of labor?"
"Is it safe to exercise while pregnant?"
"How can I prepare for breastfeeding?"

Keep answers short and consice.
"""
chat_history = []
system_message = SystemMessage(content=system_prompt)

chat_history.append(system_message)

while True:
    query = input("query: ")
    if query.lower() == "break":
        break
    chat_history.append(HumanMessage(content=query))
    result = llm.invoke(chat_history)
    response = result.content
    chat_history.append(AIMessage(content=response))
    print("AI: ", response)




